fix(codecheck): fix codecheck warnings for code quality

+++ akg-mlir/lib/Analysis/SymbolicShapeAnalysis.cpp
@@ -130,10 +130,10 @@ Type SymbolicShapeAnalysis::createNewSymbolicShape(Type type) {
   if (dict && dict.contains(getSymbolShapeAttrName())) {
     return type;
   }
-  uint64_t rank = cast<ShapedType>(type).getRank();
+  int64_t rank = cast<ShapedType>(type).getRank();
   ArrayRef<int64_t> shape = cast<ShapedType>(type).getShape();
   llvm::SmallVector<Attribute> symShapeAttr;
-  for (uint i = 0; i < rank; i++) {
+  for (int64_t i = 0; i < rank; i++) {
     if (shape[i] == ShapedType::kDynamic) {
       (void)symShapeAttr.emplace_back(StringAttr::get(type.getContext(), newSymbolicDim()));
     } else {
